A simple ruby render
At the end of the file there is a example of how it works. The
implemented primitives are Point, Line and Rectangle. There are two
types of render outputs: Ascii, which is printed as symbols and Html
where in the console is printed a ready-to-use html script with the
'picture'.
Have fun!